refactor: 커뮤니티 게시글 타입을 common으로 분리 및 적용 #1480
