+++ frontend/src/pages/community/types/posts.ts
@@ -1,14 +1,4 @@
-export interface CommunityPost {
-  id: number;
-  title: string;
-  nickname: string;
-  isAnonymous: boolean;
-  createdAt: string;
-  commentCount: number;
-  viewCount: number;
-  likeCount: number;
-  content: string;
-}
+import type { CommunityPost } from '../../../common/types/communityPost';
 
 export interface CommunityPostResponse {
   posts: CommunityPost[];
