--- frontend/src/pages/communityPostDetail/types/communityPostDetail.ts
@@ -1,12 +1,0 @@
-export interface CommunityPostDetail {
-  id: number;
-  title: string;
-  content: string;
-  nickname: string;
-  isAnonymous: boolean;
-  isGuestPost: boolean;
-  createdAt: string;
-  viewCount: number;
-  likeCount: number;
-  commentCount: number;
-}
